refactor: extract weather feature model
